requirements.txt
metaflow==2.7.4

merlin-core==0.5.0
merlin-models==0.6.0
merlin-systems==0.4.0
nvtabular==1.3.3

tensorflow-macos==2.9.2

scipy==1.9.0
tritonclient==2.24.0
